+++ fiberplane-api-client/src/lib.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::too_many_arguments)]
 #![forbid(unsafe_code)]
 #![allow(unused_mut)]
 #![allow(unused_variables)]
@@ -16,6 +17,7 @@ pub(crate) mod models {
     pub(crate) use fiberplane_models::events::*;
     pub(crate) use fiberplane_models::files::*;
     pub(crate) use fiberplane_models::formatting::*;
+    pub(crate) use fiberplane_models::front_matter_schemas::*;
     pub(crate) use fiberplane_models::labels::*;
     pub(crate) use fiberplane_models::names::*;
     pub(crate) use fiberplane_models::notebooks::operations::*;
@@ -973,6 +975,61 @@ pub async fn event_create(
         ),
     )?;
     builder = builder.json(&payload);
+    let response = builder.send().await?.error_for_status()?.json().await?;
+
+    Ok(response)
+}
+
+#[doc = r#"Retrieve all front matter schemas defined in the workspace"#]
+pub async fn workspace_front_matter_schemas_get(
+    client: &ApiClient,
+    workspace_id: base64uuid::Base64Uuid,
+) -> Result<models::WorkspaceFrontMatterSchemas> {
+    let mut builder = client.request(
+        Method::GET,
+        &format!(
+            "/api/workspaces/{workspace_id}/front_matter_schemas",
+            workspace_id = workspace_id,
+        ),
+    )?;
+    let response = builder.send().await?.error_for_status()?.json().await?;
+
+    Ok(response)
+}
+
+#[doc = r#"Retrieve a front matter schema defined in the workspace"#]
+pub async fn workspace_front_matter_schemas_create(
+    client: &ApiClient,
+    workspace_id: base64uuid::Base64Uuid,
+    payload: models::NewWorkspaceFrontMatterSchema,
+) -> Result<()> {
+    let mut builder = client.request(
+        Method::POST,
+        &format!(
+            "/api/workspaces/{workspace_id}/front_matter_schemas",
+            workspace_id = workspace_id,
+        ),
+    )?;
+    builder = builder.json(&payload);
+    let response = builder.send().await?.error_for_status()?;
+
+    Ok(())
+}
+
+#[doc = r#"Retrieve a front matter schema defined in the workspace"#]
+pub async fn workspace_front_matter_schemas_get_by_name(
+    client: &ApiClient,
+    workspace_id: base64uuid::Base64Uuid,
+    front_matter_schema_name: &str,
+) -> Result<models::FrontMatterSchema> {
+    let mut builder = client.request(
+        Method::GET,
+        &format!(
+            "/api/workspaces/{workspace_id}/front_matter_schemas/{front_matter_schema_name}",
+            workspace_id = workspace_id,
+            front_matter_schema_name = front_matter_schema_name,
+        ),
+    )?;
     let response = builder.send().await?.error_for_status()?.json().await?;
 
     Ok(response)