+++ xtask/src/commands/generate_api_client.rs
@@ -21,6 +21,7 @@ pub fn handle_generate_api_client_command() -> TaskResult {
             "fiberplane_models::events::*".to_owned(),
             "fiberplane_models::files::*".to_owned(),
             "fiberplane_models::formatting::*".to_owned(),
+            "fiberplane_models::front_matter_schemas::*".to_owned(),
             "fiberplane_models::labels::*".to_owned(),
             "fiberplane_models::names::*".to_owned(),
             "fiberplane_models::proxies::*".to_owned(),